+++ examples/agent-network/src/mastra/network/v-next.ts
@@ -1,10 +1,13 @@
 import { openai } from '@ai-sdk/openai';
+import { anthropic } from '@ai-sdk/anthropic';
 import { NewAgentNetwork } from '@mastra/core/network/vNext';
 import { Agent } from '@mastra/core/agent';
 import { createStep, createWorkflow } from '@mastra/core/workflows';
 import { LibSQLStore } from '@mastra/libsql';
 import { Memory } from '@mastra/memory';
 import { z } from 'zod';
+import { weatherAgent } from '../agents';
+import { weatherTool } from '../tools';
 
 const memory = new Memory({
   storage: new LibSQLStore({
@@ -18,7 +21,7 @@ const agent1 = new Agent({
     'This agent is used to do research, but not create full responses. Answer in bullet points only and be concise.',
   description:
     'This agent is used to do research, but not create full responses. Answer in bullet points only and be concise.',
-  model: openai('gpt-4o'),
+  model: anthropic('claude-3-5-sonnet-20240620'),
 });
 
 const agent2 = new Agent({
@@ -26,7 +29,7 @@ const agent2 = new Agent({
   description: 'This agent is used to do text synthesis on researched material. It writes articles in full paragraphs.',
   instructions:
     'This agent is used to do text synthesis on researched material. Write a full report based on the researched material. Do not use bullet points. Write full paragraphs. There should not be a single bullet point in the final report. You write articles.',
-  model: openai('gpt-4o'),
+  model: anthropic('claude-3-5-sonnet-20240620'),
 });
 
 const agentStep1 = createStep({
@@ -87,11 +90,15 @@ export const v_nextNetwork = new NewAgentNetwork({
   id: 'test-network',
   name: 'Test Network',
   instructions:
-    'You can research cities. You can also synthesize research material. You can also write a full report based on the researched material.',
-  model: openai('gpt-4o'),
+    'You can research cities. You can also synthesize research material. You can also write a full report based on the researched material. You can also get weather information. workflow1 is the best primitive for researching an *individual* city and it should always be used when running multiple primitives to accomplish a task.',
+  model: anthropic('claude-3-5-sonnet-20240620'),
   agents: {
     agent1,
     agent2,
+    // weatherAgent,
+  },
+  tools: {
+    weatherTool,
   },
   workflows: {
     workflow1,